Start: Add tests and move utils to utils/anaconda.py
