+++ spyder_kernels/console/start.py
@@ -12,15 +12,13 @@ File used to start kernels for the IPython Console
 
 # Standard library imports
 from distutils.version import LooseVersion
-import json
 import os
 import os.path as osp
-from subprocess import check_output
 import sys
 import site
 
-# Local imports
-from spyder_kernels.py3compat import to_text_string
+# Local library imports
+from spyder_kernels.utils.anaconda import is_anaconda, get_win_conda_envpath
 
 
 PY2 = sys.version[0] == '2'
@@ -78,56 +76,6 @@ init_session()
 """
 
     return lines
-
-
-def is_anaconda():
-    """
-    Detect if we are running under Anaconda.
-    Taken from https://stackoverflow.com/a/47610844/438386
-    """
-    is_conda = osp.exists(osp.join(sys.prefix, 'conda-meta'))
-    return is_conda
-
-
-def get_conda_rootpath():
-    """Get conda root path using 'conda info --json'."""
-    try:
-        conda_info = check_output('conda info --json')
-        conda_info_json = json.loads(conda_info)
-        return to_text_string(conda_info_json['env_vars']['CONDA_ROOT'])
-    except Exception:
-        return None
-
-
-def get_win_conda_envpath():
-    """
-    Get minimal set of directories to add to PATH for conda
-    environments on Windows.
-    """
-    env_python_exec = sys.executable
-    conda_root_path = get_conda_rootpath()
-
-    # Environment PATH elements
-    env_path = ''
-    if env_python_exec:
-        env_root_path = osp.dirname(to_text_string(env_python_exec))
-        env_path = env_root_path + os.pathsep
-        env_path += osp.join(env_root_path, u'Library', u'mingw-w64',
-                             u'bin' + os.pathsep)
-        env_path += osp.join(env_root_path, u'Library', u'usr',
-                             u'bin' + os.pathsep)
-        env_path += osp.join(env_root_path, u'Library', u'bin' + os.pathsep)
-        env_path += osp.join(env_root_path, u'Scripts' + os.pathsep)
-        env_path += osp.join(env_root_path, u'bin' + os.pathsep)
-
-    # Root environment PATH elements
-    root_path = ''
-    if conda_root_path is not None:
-        root_path = conda_root_path + ';'
-        env_path += osp.join(conda_root_path, u'Scripts' + os.pathsep)
-        env_path += osp.join(conda_root_path, u'bin' + os.pathsep)
-
-    return env_path + root_path
 
 
 def kernel_config():
@@ -348,7 +296,9 @@ def main():
     while '' in sys.path:
         sys.path.remove('')
 
-    # Handle conda environment. See Spyder issue #9077
+    # Add to PATH most of the directories added when activating a conda
+    # environment.
+    # See https://github.com/spyder-ide/spyder/issues/9077
     if os.environ.get('SPY_EXTERNAL_INTERPRETER') and os.name == 'nt':
         if is_anaconda():
             conda_env_path = get_win_conda_envpath()
